requirements.txt
langchain
sentence-transformers
faiss-cpu
pinecone-client[grpc]
pdfminer.six
pdf2image
pytesseract
unstructured[ocr]
python-dotenv
openai
uvicorn
fastapi
pydantic
torch
transformers
langdetect
deep-translator
gtts
pygame
sounddevice
langchain-openai
unstructured[ocr]
unstructured-inference
sentence-transformers
pinecone-client
openai
python-dotenv
